+++ plugins/scaffolder-react/src/next/components/TemplateCard/CardLink.tsx
