+++ plugins/scaffolder-react/src/next/components/TemplateCard/TemplateCard.test.tsx
@@ -25,9 +25,9 @@ import {
 } from '@backstage/test-utils';
 import { TemplateCard } from './TemplateCard';
 import React from 'react';
-import { nextRouteRef } from '@backstage/plugin-scaffolder-react';
 import { TemplateEntityV1beta3 } from '@backstage/plugin-scaffolder-common';
 import { RELATION_OWNED_BY } from '@backstage/catalog-model';
+import { nextRouteRef } from '../../../routes';
 
 describe('TemplateCard', () => {
   it('should render the card title', async () => {
