+++ plugins/scaffolder-react/src/next/components/TemplateCard/index.ts
