+++ plugins/scaffolder-react/src/next/index.ts
@@ -15,3 +15,5 @@
  */
 export * from './components';
 export * from './extensions';
+export * from './types';
+export * from './lib';
