+++ plugins/scaffolder-react/src/next/lib/index.ts
@@ -13,6 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from './Stepper';
-export * from './TemplateCard';
-export * from './ReviewState';
+export { extractSchemaFromStep, createFieldValidation } from './schema';
