+++ plugins/scaffolder-react/src/next/lib/schema.test.ts
