+++ plugins/scaffolder-react/src/next/lib/schema.ts
