+++ plugins/scaffolder/src/components/TemplateEditorPage/TemplateEditorForm.tsx
@@ -19,9 +19,11 @@ import { makeStyles } from '@material-ui/core/styles';
 import React, { Component, ReactNode, useMemo, useState } from 'react';
 import useDebounce from 'react-use/lib/useDebounce';
 import yaml from 'yaml';
-import { FieldExtensionOptions } from '../../extensions';
+import {
+  FieldExtensionOptions,
+  TemplateParameterSchema,
+} from '@backstage/plugin-scaffolder-react';
 import { LayoutOptions } from '../../layouts';
-import { TemplateParameterSchema } from '../../types';
 import { MultistepJsonForm } from '../MultistepJsonForm';
 import { createValidator } from '../TemplatePage';
 import { useDirectoryEditor } from './DirectoryEditorContext';
