+++ plugins/scaffolder/src/next/TemplateWizardPage/TemplateWizardPage.tsx
@@ -22,7 +22,6 @@ import {
   InfoCard,
   MarkdownContent,
 } from '@backstage/core-components';
-import { NextFieldExtensionOptions } from '../../extensions';
 import { Navigate, useNavigate } from 'react-router-dom';
 import { stringifyEntityRef } from '@backstage/catalog-model';
 import {
@@ -35,7 +34,10 @@ import {
 import { scaffolderApiRef } from '../../api';
 import useAsync from 'react-use/lib/useAsync';
 import { makeStyles } from '@material-ui/core';
-import { Stepper } from '@backstage/plugin-scaffolder-react';
+import {
+  Stepper,
+  NextFieldExtensionOptions,
+} from '@backstage/plugin-scaffolder-react';
 import { BackstageTheme } from '@backstage/theme';
 import {
   nextRouteRef,
@@ -44,7 +46,7 @@ import {
 } from '@backstage/plugin-scaffolder-react';
 import { SecretsContext } from '../../components/secrets/SecretsContext';
 import { JsonValue } from '@backstage/types';
-import type { FormProps } from '../types';
+import type { FormProps } from '@backstage/plugin-scaffolder-react';
 
 export type TemplateWizardPageProps = {
   customFieldExtensions: NextFieldExtensionOptions<any, any>[];
